+++ test/AdventOfCode.Tests/2023/Day05/SeedConverterTest.cs
@@ -33,18 +33,18 @@ public class SeedConverterTest
 
 public class SeedConverter
 {
-    private readonly int destinationRangeStart;
-    private readonly int rangeLength;
-    private readonly int sourceRangeStart;
+    private readonly long destinationRangeStart;
+    private readonly long rangeLength;
+    private readonly long sourceRangeStart;
 
-    public SeedConverter(int destinationRangeStart, int sourceRangeStart, int rangeLength)
+    public SeedConverter(long destinationRangeStart, long sourceRangeStart, long rangeLength)
     {
         this.destinationRangeStart = destinationRangeStart;
         this.sourceRangeStart = sourceRangeStart;
         this.rangeLength = rangeLength;
     }
 
-    public int GetDestination(int source)
+    public long GetDestination(long source)
     {
         if (IsInRange(source) == false)
             return source;
@@ -52,7 +52,7 @@ public class SeedConverter
         return destinationRangeStart + (source - sourceRangeStart);
     }
 
-    public bool IsInRange(int source)
+    public bool IsInRange(long source)
         => source >= sourceRangeStart &&
            source < sourceRangeStart + rangeLength;
 }
